requirements/base.txt
certifi==2017.7.27.1
chardet==3.0.4
click==6.7
coreapi==2.3.3
coreschema==0.0.4
Django==2.0.3
django-cleanup==2.1.0
django-extensions==2.0.7
django-model-utils==3.0.0
django-modeladmin-reorder==0.3
django-oauth-toolkit==1.1.0
django-rest-swagger==2.1.2
djangorestframework==3.8.2
drf-yasg==1.7.4
flex==6.13.2
future==0.16.0
idna==2.6
inflection==0.3.1
itypes==1.1.0
Jinja2==2.9.6
jsonpointer==1.14
jsonschema==2.6.0
MarkupSafe
numpy==1.14.3
oauthlib==2.0.7
olefile==0.44
openapi-codec==1.3.2
pandas==0.23.3
Pillow==4.3.0
psycopg2==2.7.3.2
python-dateutil==2.7.3
pytz==2018.3
PyYAML==3.12
requests==2.18.4
rfc3987==1.3.7
ruamel.yaml==0.15.37
simplejson==3.11.1
six==1.11.0
strict-rfc3339==0.7
stripe==1.79.1
swagger-spec-validator==2.1.0
uritemplate==3.0.0
urllib3==1.22
validate-email==1.3
django-admin-rangefilter==0.3.10
pycountry==18.5.26
ObjDict==0.4.4
scipy==1.3.1
linux-metrics==0.1.4
selenium==3.141.0
